+++ src/xslint.js
@@ -16,6 +16,8 @@ const {lintByAxis, names: axisChecks} = require('./xpath-axis-linter')
 const {lintByNamespace, names: namespaceChecks} = require('./namespace-linter')
 const {lintByNodeSet, names: nodeSetChecks} = require('./node-set-linter')
 const {lintByCount, names: countChecks} = require('./count-linter')
+const {lintByStringLength, names: stringLengthChecks} =
+  require('./string-length-linter')
 const {lintByFormat, names: formatChecks} = require('./xpath-format-linter')
 const {fixed} = require('./fixer')
 const {logger, levels} = require('./logger')
@@ -36,6 +38,7 @@ const LINTERS = [
   lintByNamespace,
   lintByNodeSet,
   lintByCount,
+  lintByStringLength,
 ]
 
 /**
@@ -56,7 +59,7 @@ const EXPRESSION_LINTERS = [
 const CHECKS = [
   ...xslChecks, ...xpathValidatorChecks,
   ...xpathChecks, ...corpusChecks, ...axisChecks, ...namespaceChecks,
-  ...nodeSetChecks, ...countChecks, ...formatChecks,
+  ...nodeSetChecks, ...countChecks, ...stringLengthChecks, ...formatChecks,
 ]
 
 /**