+++ test/xcop.test.js
@@ -22,6 +22,7 @@ const PACKS = [
   ...allFilesFrom(path.resolve(__dirname, 'resources', 'namespace-packs')),
   ...allFilesFrom(path.resolve(__dirname, 'resources', 'node-set-packs')),
   ...allFilesFrom(path.resolve(__dirname, 'resources', 'count-packs')),
+  ...allFilesFrom(path.resolve(__dirname, 'resources', 'string-length-packs')),
 ]
 
 /**